This is Uche committing on Michau's behalf.
He added a new tab to the Shuffleboard to store data coming from the
InputSubsystem.  And if that sounds familiar, that's because it should
be: we did the exact same thing with the test-bed a few weeks ago.

Right now, his changes record ForwardBack, LeftRight, and Rotation in
three widgets under the "Input" tab.  You may need to drag that tab in
from the LiveWindow, as the tab doesn't seem to be showing up in the
ShuffleBoard at the moment.  It's there, though: trust me.

Next step?  Figure out what values we're passing into the actual
Mecanum drive -- and perhaps use the REV Hardware Client to debug
individual motors.