This is Uche, committing on Maxwell's behalf.
We were working on PID control for the ArmSubsystem when we realized
that we wanted the ability to read from AND write to ShuffleBoard
values.  The reading happens when the user changes the shuffleboard and
wishes for us to reflect that; the writing happens when we calculate
a new value for the arm angles and need to show this to the user.

How do we tell if a widget's underlying value just changed so as to
update it?  We're not sure yet; committing anyway.